Modernize Joomla site/civicrm.php
This was triggered by debugging why you could not have event listing
on the front page in Joomla 4. It does away with "backward
compatibility" for Joomla 2.x (reading $_REQUEST) in favour of J3 (and
up) Input method.

It also removes some unused variables and code, and replaces old
JFactory with Factory.