(dev/core#5128) Installer/Upgrader misfires in presence of certain passwords
Overview
--------

CiviCRM-Joomla upgrade fails if (1) the MySQL password includes a URL
meta-character (like `#`) and (2) if the -prior- version of CiviCRM was 5.66
or older.

Specifically, the failure relates to `civicrm_detect_upgrade()`.  This
utility function analyzes your MySQL setup and decides whether to do an
installation or upgrade.  If the function misfires, then you wind up
executing a new-installation instead of executing the upgrader.

Steps to Reproduce
------------------

* Configure MySQL with a password like `a#a{`. Note that `#` is a URL metacharacter.
* Install Joomla 4.4.3
* Install CiviCRM 5.66.2. (This writes a `CIVICRM_DSN` with a password containing literal `#`, which not URL-conformant.)
* Upgrade to CiviCRM 5.73.3

Before
------

`civicrm_detect_upgrade()` examines the MySQL settings.  It uses
`parse_url()` to inspect the old value of `CIVICRM_DSN`.  The `CIVICRM_DSN`
includes a literal `#` in the password. `parse_url()` interprets this
as a fragment -- and then fails to identify the database-name.

After
-----

`civicrm_detect_upgrade()` examines the MySQL settings.  It uses
`DB::parseDSN()`, which is more generous about accepting non-conformant
URLs.  (It accepts a password with literal `#`.)

Comments
--------

* In CiviCRM `<=5.66`, the `CIVICRM_DSN` is written with non-URL-conformant notation.
* In CiviCRM `>=5.67`, the `CIVICRM_DSN` is written with URL-conformant notation.
* In CiviCRM `5.72`, the installer-upgrader starts to use `parse_url()`; it
  assumes URL-conformance.